Started working on proper contao 4 symfony bundle
