--- config/autoload.ini
@@ -1,14 +1,0 @@
-
-;;
-; List modules which are required to be loaded beforehand
-;;
-requires[] = "core"
-requires[] = "haste"
-requires[] = "multicolumnwizard"
-
-;;
-; Configure what you want the autoload creator to register
-;;
-register_namespaces = false
-register_classes    = false
-register_templates  = false
