+++ src/DataTransformer/AbstractDateTransformer.php
@@ -8,7 +8,8 @@
  * @license    http://opensource.org/licenses/lgpl-3.0.html LGPL
  * @link       http://github.com/terminal42/contao-leads
  */
-namespace Leads\DataTransformer;
+
+namespace Terminal42\Leads\DataTransformer;
 
 
 abstract class AbstractDateTransformer implements DataTransformerInterface, DisplayInBackendInterface
