+++ src/DataTransformer/DatimTransformer.php
@@ -8,8 +8,8 @@
  * @license    http://opensource.org/licenses/lgpl-3.0.html LGPL
  * @link       http://github.com/terminal42/contao-leads
  */
-namespace Leads\DataTransformer;
 
+namespace Terminal42\Leads\DataTransformer;
 
 class DatimTransformer extends AbstractDateTransformer
 {
