+++ src/Exporter/ExporterInterface.php
@@ -8,18 +8,23 @@
  * @license    http://opensource.org/licenses/lgpl-3.0.html LGPL
  * @link       http://github.com/terminal42/contao-leads
  */
-namespace Leads\Exporter;
+namespace Terminal42\Leads\Exporter;
 
-class Xls extends AbstractExcelExporter
-{
+
+interface ExporterInterface {
+
+    /**
+     * Returns true if available.
+     *
+     * @return bool
+     */
+    public function isAvailable();
+
     /**
      * Exports a given set of data row ids using a given configuration.
      *
      * @param \Database\Result $config
      * @param array|null       $ids
      */
-    public function export($config, $ids = null)
-    {
-        $this->exportWithFormat($config, $ids, 'Excel5');
-    }
+    public function export($config, $ids = null);
 }
