+++ src/Exporter/Utils/Tokens.php
@@ -9,10 +9,9 @@
  * @link       http://github.com/terminal42/contao-leads
  */
 
-namespace Leads\Exporter\Utils;
+namespace Terminal42\Leads\Exporter\Utils;
 
-
-use Leads\DataCollector;
+use Terminal42\Leads\DataCollector;
 
 class Tokens
 {
