+++ src/Exporter/Xls.php
@@ -8,23 +8,18 @@
  * @license    http://opensource.org/licenses/lgpl-3.0.html LGPL
  * @link       http://github.com/terminal42/contao-leads
  */
-namespace Leads\Exporter;
+namespace Terminal42\Leads\Exporter;
 
-
-interface ExporterInterface {
-
-    /**
-     * Returns true if available.
-     *
-     * @return bool
-     */
-    public function isAvailable();
-
+class Xls extends AbstractExcelExporter
+{
     /**
      * Exports a given set of data row ids using a given configuration.
      *
      * @param \Database\Result $config
      * @param array|null       $ids
      */
-    public function export($config, $ids = null);
+    public function export($config, $ids = null)
+    {
+        $this->exportWithFormat($config, $ids, 'Excel5');
+    }
 }
