+++ src/Resources/contao/config/config.php
