+++ src/Resources/contao/languages/de/modules.php
