+++ src/Resources/contao/languages/fr/modules.php
