+++ src/Resources/contao/languages/fr/tl_form.php
